Make edits for CheckStyle test
