+++ src/main/java/seedu/address/logic/cap/Logic.java
@@ -3,14 +3,13 @@ package seedu.address.logic.cap;
 import java.nio.file.Path;
 
 import javafx.collections.ObservableList;
-import javafx.scene.image.Image;
 import javafx.scene.chart.PieChart;
+import javafx.scene.image.Image;
 import seedu.address.commons.core.GuiSettings;
 import seedu.address.logic.cap.commands.CommandResult;
 import seedu.address.logic.cap.commands.exceptions.CommandException;
 import seedu.address.logic.cap.parser.exceptions.ParseException;
 import seedu.address.model.cap.ReadOnlyCapLog;
-import seedu.address.model.cap.person.Rank.Rank;
 import seedu.address.model.cap.person.Semester;
 import seedu.address.model.common.Module;
