+++ src/main/java/seedu/address/model/cap/Model.java
@@ -7,7 +7,6 @@ import javafx.collections.ObservableList;
 import javafx.scene.chart.PieChart;
 import javafx.scene.image.Image;
 import seedu.address.commons.core.GuiSettings;
-import seedu.address.model.cap.person.Rank.Rank;
 import seedu.address.model.cap.person.Semester;
 import seedu.address.model.common.Module;
