+++ src/main/java/seedu/address/model/cap/person/rank/FirstClass.java
@@ -1,7 +1,10 @@
-package seedu.address.model.cap.person.Rank;
+package seedu.address.model.cap.person.rank;
 
 import seedu.address.model.cap.person.DegreeClassification;
 
+/**
+ * Represents a FirstClass grade.
+ */
 public class FirstClass extends Rank {
 
     private static final String TITLE = "FIRST CLASS HONOURS";
@@ -26,6 +29,11 @@ public class FirstClass extends Rank {
         return imageFilePath;
     }
 
+    /**
+     * To check if the cap given is within the range of the catagorisation of this rank/ class.
+     * @param cap CUMULATIVE AVERAGE POINT
+     * @return
+     */
     public static boolean isWithinRange(double cap) {
         if (cap >= MINIMUM_CAP && cap <= MAXIMUM_CAP) {
             return true;
