+++ src/main/java/seedu/address/model/cap/person/rank/SecondLower.java
@@ -1,7 +1,10 @@
-package seedu.address.model.cap.person.Rank;
+package seedu.address.model.cap.person.rank;
 
 import seedu.address.model.cap.person.DegreeClassification;
 
+/**
+ * Represents a SecondLower grade.
+ */
 public class SecondLower extends Rank {
 
     private static final String TITLE = "SECOND CLASS HONOURS (LOWER)";
@@ -30,6 +33,11 @@ public class SecondLower extends Rank {
         return imageFilePath;
     }
 
+    /**
+     * To check if the cap given is within the range of the catagorisation of this rank/ class.
+     * @param cap CUMULATIVE AVERAGE POINT
+     * @return
+     */
     public static boolean isWithinRange(double cap) {
         if (cap >= MINIMUM_CAP && cap <= MAXIMUM_CAP) {
             return true;
