+++ src/main/java/seedu/address/model/cap/person/rank/ThirdClass.java
@@ -1,7 +1,10 @@
-package seedu.address.model.cap.person.Rank;
+package seedu.address.model.cap.person.rank;
 
 import seedu.address.model.cap.person.DegreeClassification;
 
+/**
+ * Represents a ThirdClass grade.
+ */
 public class ThirdClass extends Rank {
 
     private static final String TITLE = "THIRD CLASS HONOURS";
@@ -26,6 +29,11 @@ public class ThirdClass extends Rank {
         return imageFilePath;
     }
 
+    /**
+     * To check if the cap given is within the range of the catagorisation of this rank/ class.
+     * @param cap CUMULATIVE AVERAGE POINT
+     * @return
+     */
     public static boolean isWithinRange(double cap) {
         if (cap >= MINIMUM_CAP && cap <= MAXIMUM_CAP) {
             return true;
